move file to schema
